🔥Del : 사용하지 않는 CellView와 CellViewModel 삭제
