--- ProjectManager/ViewModel/CellViewModel.swift
@@ -1,27 +1,0 @@
-//
-//  CellViewModel.swift
-//  ProjectManager
-//
-//  Created by OneTool, marisol on 2022/07/20.
-//
-
-import SwiftUI
-
-class CellViewModel: ViewModelType {
-    @Published var isShowingSheet = false
-    @Published var isShowingPopover = false
-    @Published var task: Task
-    
-    init(withService: TaskManagementService, task: Task) {
-        self.task = task
-        super.init(withService: withService)
-    }
-    
-    func toggleShowingSheet() {
-        isShowingSheet.toggle()
-    }
-    
-    func toggleShowingPopover() {
-        isShowingPopover.toggle()
-    }
-}
